backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
openai
Pillow
numpy
pyzbar
